extended ukf for position dimensions

+++ src/modules/quat/utils/quat_constants.h
@@ -14,6 +14,6 @@
 #define COMPASS_DECLINATION		2.36888889f//East in degrees//2.2666f//Declination +Ost-West   -10.11f	// local magnetic compass correction to true north [Cove]
 #define COMPASS_INCLINATION		-64.132470167481f
 #define RAD_TO_DEG		(180.0f/M_PI)
-#define DEG_TO_RAD		(M_PI/180.0f)
+#define DEG_TO_RAD		((float)M_PI/180.0f)
 
 #endif /* QUAT_CONSTANTS_H_ */
